// frontend/src/types/generated/MessageWithRun.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RunWithMeta } from "./RunWithMeta";

export type MessageWithRun = { id: bigint, task_id: number, role: string, content: string, created_at: string | null, metadata: Record<string, any> | null, run: RunWithMeta | null, };